Encoding array params correctly in toString method
This PR is related to this issue: https://github.com/angular/angular/issues/19071
The change allows for correctly encoding array params. Values is in this method always an array, but when length is more then 1 it means we have several values from an object that has been passed to the encoder. These values need to be encoded using '[]' so a backend understands we have an array of values,

Example:

    const object = { key: ['foo','bar'] };
    const httpParams = new HttpParams({
       fromObject:  object,
    });

Previous result before PR: 

    key=foo&key=bar

New result after PR:

    key[]=foo&key[]=bar